% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doubleTree.R
\name{doubleTree}
\alias{doubleTree}
\title{Function to plot multidimensional data}
\usage{
doubleTree(
  x,
  y,
  z1,
  z2 = NULL,
  c1 = NULL,
  c2 = NULL,
  c3 = NULL,
  c4 = NULL,
  mid = 1/3,
  xscale = NULL,
  yscale = NULL,
  zscale = NULL,
  xar = 0.03,
  grid.col = "gray",
  lab.col = "gray20",
  lab.cex = 0.7,
  lab.dig = 1,
  lab.dig.x = lab.dig,
  lab.dig.y = lab.dig,
  lab.dig.z = lab.dig,
  vert.lines = T,
  vert.wide = 1,
  vert.anot = F,
  vert.col = "black",
  xlab = "X",
  ylab = "Y",
  zlab = "Z",
  direct.leg = T,
  direct.leg.x = -0.9,
  direct.leg.y = 0.95,
  arrow.leg = T,
  arrow.txt = "",
  rev.palette = F,
  c1.col = NULL,
  c2.col = NULL,
  c3.col = NULL,
  c4.col = NULL
)
}
\arguments{
\item{x}{numeric variable, NA not allowed}

\item{y}{numeric variable, NA not allowed}

\item{z1}{numeric variable specifies the tip of the arrow head}

\item{z2}{if sepecifies the base of the arrow head (default NULL)}

\item{c1}{variable for the color of the left part of the top arrow head (dafault = z1)}

\item{c2}{variable for the color of the right part of the top arrow head (dafault = c1)}

\item{c3}{variable for the color of the left part of the bottom arrow head (dafault = z1)}

\item{c4}{variable for the color of the right part of the bottom arrow head (dafault = c1)}

\item{mid}{specifying the ration of the base grid to the Z expansion (default = 1/3)}

\item{xscale}{Manual scale for x, usually you should use the default (auto)}

\item{xar}{Width of the arrow base (default = 0.03)}

\item{lab.dig}{lab.dig.x ... number of digits for the x/y grid annotation}

\item{xlab}{ylab zlab axis labels (similar to base plot)}

\item{direct.leg}{Should arrow directions shown as legend (default = F)}

\item{arrow.leg}{Should an arrow color legend be added (default= F)}

\item{rev.palette}{Should the default color palette reverted that red is low and green = high?}

\item{c1.col}{color palette for c1 (default from green to red)}

\item{c2.col}{color palette for c2 (default from green to red)}

\item{c3.col}{color palette for c3 (default from green to red)}

\item{c4.col}{color palette for c4 (default from green to red)}
}
\value{
invisible(NULL)
}
\description{
Function to plot highly multidemensional data.
Variables x, y, and z1 are plotted similar to a 3d scatter plot
except that in z dimension an arrow is plotted.
The double arrow head has either a fixed size or - if z1 and z2 ar both provided -
the head size ranges from z2 to z1.
The heads are showing either up or down if z1-z2 is pos or negative.
The head can have additional color codes c1 to c4.
The plot can be quite complex. Best way is to check the examples.
}
\examples{
doubleTree(x = iris$Sepal.Width,
           y = iris$Sepal.Length,
           z1 = iris$Petal.Width,
           z2 = iris$Petal.Length,
           c1 = as.numeric(iris$Species),
           c3 = iris$Petal.Width,
           c4 = iris$Petal.Length,
           xlab = "Sepal.Width",
           ylab = "Sepal.Length",
           zlab="Petal")

df <- data.frame(x = cos(seq(0, pi*1.3, length.out=20)),
                 y = sin(seq(0, pi*1.3, length.out=20)),
                 z1 = 1:20,
                 z2 = seq(15, 5, length.out=20),
                 c1 = rep(1:3, 7)[1:20],
                 c2 = rep(4:1, 6)[1:20],
                 c3 = rep(0:1, each = 10),
                 c4 = 1:20)

 doubleTree(x = df$x,
            y = df$y,
            z1 = df$z1,
            z2 = df$z2,
            c1 = df$c1,
            c2 = df$c2,
            c3 = df$c3,
            c4 = df$c4,
            c1.col = c("blue","cornflowerblue","orchid"),
            c2.col = c("darkgreen","green","lightgreen"),
            c3.col = c("black","white"),
            arrow.leg = T,
            xar=0.05)

}
