% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmastree.R
\name{xmastree}
\alias{xmastree}
\title{Generate a Xmas Tree}
\usage{
xmastree()
}
\value{
a ggplot2 object
}
\description{
Simple function that generates a unique christmas tree, varying
the colors based on the uniform distribution
}
